+++ accomodation/src/main/java/com/codecool/accomodation/model/entity/Host.java
@@ -1,4 +1,4 @@
-package com.codecool.accomodation.entity;
+package com.codecool.accomodation.model.entity;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import lombok.AllArgsConstructor;
@@ -7,7 +7,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.util.List;
 import java.util.Set;
 
 @Data
