+++ accomodation/src/main/java/com/codecool/accomodation/model/entity/types/AccommodationType.java
@@ -1,4 +1,4 @@
-package com.codecool.accomodation.entity.types;
+package com.codecool.accomodation.model.entity.types;
 
 public enum AccommodationType {
