+++ accomodation/src/main/java/com/codecool/accomodation/repository/HostRepository.java
@@ -1,6 +1,6 @@
 package com.codecool.accomodation.repository;
 
-import com.codecool.accomodation.entity.Host;
+import com.codecool.accomodation.model.entity.Host;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HostRepository extends JpaRepository<Host, Long> {
