+++ accomodation/src/main/java/com/codecool/accomodation/repository/LocationRepository.java
@@ -1,6 +1,6 @@
 package com.codecool.accomodation.repository;
 
-import com.codecool.accomodation.entity.Location;
+import com.codecool.accomodation.model.entity.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
