+++ accomodation/src/main/java/com/codecool/accomodation/repository/RoomRepository.java
@@ -1,6 +1,6 @@
 package com.codecool.accomodation.repository;
 
-import com.codecool.accomodation.entity.Room;
+import com.codecool.accomodation.model.entity.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomRepository extends JpaRepository<Room, Long> {
